requirements.txt
numpy
pandas
torch
arxiv
crewai>=0.12.1
crewai-tools
google-generativeai
docx
requests
transformers
ctransformers
sentence-transformers
streamlit-feedback
streamlit
lmstudio
reportlab
python-dotenv
pydantic>=2.0.0
openai
langchain
langchain-community
langchain_google_genai
# fpdf
# PyPDF2
# pdfkit
# chromadb
# mistral_inference
# huggingface_hub

# ollama

